requirements.txt
aiohappyeyeballs==2.3.2
aiohttp==3.10.0
aiosignal==1.3.1
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
black==24.4.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==5.0.7
djangorestframework==3.15.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
multidict==6.0.5
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
python-dotenv==1.0.1
python-telegram-bot==21.4
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
yarl==1.9.4